requirements.txt
typer>=0.12.3
rich>=13.7.1
questionary>=2.0.1
gitpython>=3.1.43
nltk>=3.8.1
torch>=2.2.0
transformers>=4.40.0
numpy>=1.26.0
sentence-transformers>=2.7.0
plotille>=5.0.0
pyfiglet>=1.0.2
